Revert "Changing parameter in LanguageFactory.cs"
This reverts commit ee737841e3b39cb2c1e530d3693421cf7e12d522.

+++ src/Umbraco.Infrastructure/Persistence/Factories/LanguageFactory.cs
@@ -30,7 +30,7 @@ internal static class LanguageFactory
         return lang;
     }
 
-    public static LanguageDto BuildDto(ILanguage entity, short? fallbackLanguageId)
+    public static LanguageDto BuildDto(ILanguage entity, int? fallbackLanguageId)
     {
         ArgumentNullException.ThrowIfNull(entity);
